+++ user-service/src/main/java/com/isloand/userservice/repository/UserRepository.java
@@ -1,6 +1,6 @@
 package com.isloand.userservice.repository;
 
-import com.isloand.userservice.domain.User;
+import com.isloand.userservice.entity.User;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface UserRepository extends JpaRepository<User, Long> {
